feature:Add Privacy Policy page detailing data handling practices
